+++ Hotel_Management_System/Hotel_Management_System/Model/PHIEUNHAPHANG.cs
@@ -23,7 +23,7 @@ namespace Hotel_Management_System.Model
         public string MaPhieuNhapHang { get; set; }
         public string MaNhanVien { get; set; }
         public Nullable<System.DateTime> NgayLap { get; set; }
-        public Nullable<double> TriGia { get; set; }
+        public Nullable<int> TriGia { get; set; }
     
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2227:CollectionPropertiesShouldBeReadOnly")]
         public virtual ICollection<CT_PHIEUNHAPHANG> CT_PHIEUNHAPHANG { get; set; }
